# dispatch-management-system/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/v1
  description: Generated server url
security:
- basicAuth: []
paths:
  /despachos:
    get:
      tags:
      - shipment-controller
      operationId: getAllShipments
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShipmentResponse"
    post:
      tags:
      - shipment-controller
      operationId: createShipment
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        description: Clave única para evitar registrar dos veces la misma solicitud.
        schema:
          type: string
          maxLength: 255
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentRequest"
        required: true
      responses:
        "201":
          description: Despacho creado o resultado previo de la misma solicitud.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShipmentResponse"
  /despachos/{id}/revertir:
    post:
      tags:
      - shipment-controller
      operationId: revertShipment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShipmentResponse"
  /clientes:
    get:
      tags:
      - client-controller
      operationId: getAllClients
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientResponse"
    post:
      tags:
      - client-controller
      operationId: createClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
        required: true
      responses:
        "201":
          description: Cliente creado.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientResponse"
  /despachos/{id}:
    get:
      tags:
      - shipment-controller
      operationId: getShipmentById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShipmentResponse"
  /clientes/{id}:
    get:
      tags:
      - client-controller
      operationId: getClientById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClientResponse"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    CreateShipmentRequest:
      type: object
      properties:
        clientId:
          type: integer
          format: int32
        truckId:
          type: integer
          format: int32
        weightKg:
          type: integer
          format: int32
          minimum: 1
        detail:
          type: string
          maxLength: 300
          minLength: 1
      required:
      - clientId
      - detail
      - truckId
      - weightKg
    ShipmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        clientId:
          type: integer
          format: int32
        truckId:
          type: integer
          format: int32
        loadId:
          type: integer
          format: int32
        weightKg:
          type: integer
          format: int32
        detail:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        _links:
          type: object
          description: Enlaces disponibles para el despacho según su estado.
          additionalProperties:
            type: object
            properties:
              href:
                type: string
                format: uri-reference
            required:
            - href
      required:
      - _links
    CreateClientRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
          minLength: 1
        email:
          type: string
          format: email
          maxLength: 150
          minLength: 1
        phone:
          type: string
          maxLength: 20
          minLength: 0
      required:
      - email
      - name
    ClientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        email:
          type: string
        phone:
          type: string